# SC2: Region access rule speedups