## What is this fixing or adding?
These changes are in response to the reported slowdown of SC2's template YAML following the recent content merge, targeting issues with the region code. 

The steps taken were:
- Replacing `sum()` calls in lambdas with explicit list iterations and early return conditions
  - In one case a function call to create strings was moved out of the containing lambda
- Special-casing small entry rules to avoid creating lambdas that just call one other lambda
- Removing a lambda that was created for every access rule and would only ever call one other lambda
- Manually preventing lambdas that would only ever evaluate to `True` from being called in the final access rules
- Implementing a simple algorithm to find a mission that *must* be beaten to satisfy an entry rule, to reduce connections to `Menu` in the specific case of vanilla-like mission orders

## How was this tested?
Using cProfile to run test generations with 3 SC2 template YAMLs. I tested this incrementally to verify each step, but for brevity I'll only showcase a full comparison below. I don't know if this known or intended, but I found playthrough calculation to be a source of non-determinism with regards to total function calls, so I've included `update_reachable_regions` calls as a measure of how "lucky" the generation was, and have otherwise stripped code I haven't touched.

Before any changes:
```
Done. Enjoy. Total Time: 136.55459279986098
         229991172 function calls (225527525 primitive calls) in 143.492 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   157818    0.466    0.000   34.274    0.000 BaseClasses.py:748(update_reachable_regions)
3509784/3363821    2.402    0.000   24.627    0.000 entry_rules.py:238(<lambda>)
6585451/6147562    2.163    0.000   20.490    0.000 entry_rules.py:238(<genexpr>)
  1317925    0.602    0.000   13.965    0.000 generation.py:510(<lambda>)
  1317925    1.110    0.000   13.364    0.000 generation.py:499(<lambda>)
   809927    0.407    0.000   12.780    0.000 generation.py:506(<lambda>)
   986610    0.738    0.000   12.342    0.000 entry_rules.py:158(<lambda>)
  6931151    4.769    0.000   10.283    0.000 entry_rules.py:158(<genexpr>)
  8555102    2.539    0.000    2.539    0.000 nodes.py:567(beat_item)
 1050/966    0.001    0.000    0.004    0.000 entry_rules.py:224(_is_fulfilled)
        3    0.001    0.000    0.004    0.001 generation.py:486(make_connections)
```
Note that the `generation.py:510` and `generation.py:506` lambdas were pointless wrappers and thus removed. The `generation.py:499` lambda still exists, but the code to filter out functionally useless lambdas from access rules prevents this lambda from being created on the vanilla mission order, so it doesn't show up in the next profile.

With all changes:
```
Done. Enjoy. Total Time: 116.06302800029516
         188236306 function calls (186245313 primitive calls) in 122.774 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   157241    0.444    0.000   13.356    0.000 BaseClasses.py:748(update_reachable_regions)
   371752    0.463    0.000    4.816    0.000 entry_rules.py:279(count_rules) -- formerly entry_rules.py:238
   955152    2.395    0.000    4.424    0.000 entry_rules.py:172(count_missions) -- formerly entry_rules.py:158
  1204884    0.403    0.000    0.403    0.000 nodes.py:567(beat_item)
        3    0.001    0.000    0.005    0.002 generation.py:486(make_connections)
 1050/966    0.001    0.000    0.003    0.000 entry_rules.py:252(_is_fulfilled)
```
At first glance there's a discrepancy between the total 20s gain and the individual gains of `count_rules` (24s -> 4s) and `count_missions` (12s -> 4s), but this can be explained. As a guarantee of the system, all access rules have a `SubRuleEntryRule` as their top-most rule, so in the original code `entry_rules.py:238` includes the run time of `entry_rules.py:158`, but in the new code `count_missions` can bypass `count_rules` via a special case, so the gains of `count_rules` include a duplicate 8s gain from never being called.

To conclude, these changes come out to a ~20s or ~14% speedup on my machine.